moved estimate_extrinsic_weight
